requirements.txt
fastapi
uvicorn
python-dotenv
pyodbc
requests
apscheduler